Assuming FAT blocks are continuous
Previously, the FAT library has to read the File Allocation Table to
determine the next block of the file. Obviously that is not cached, so
retrieving the next block address has a very high overhead. Since for
now the blocks seems to be continuous, we'll assume that until we have a
new design that uses a SPI flash and USB Mass Storage