Fix node searching bug
This bug currently affects words with default orthography, i.e. the ones
that you just add a s to it, for example. This bug is discovered with
the stroke sequence REBG/ORD/S, which should output "records", but
was outputting "orders", as there's no orthography rules for "record"
because you just need to add s to it, and the engine sees that "orders"
is the longest match, and uses it.

This is resolved by enforcing that the next non root stroke must have
one more level than the current/last output. If that is not possible,
than we simply use the root based node.