Add keycode parsing and modifier support
Keycodes used to be supported as a pair of bytes, with mod information
in one and the key code itself in the other. But that would not support
holding down the modifier across keys, so the format needs to be
changed. Now, the key code format is similar to the strings format, with
a length in front and then the key codes. The modifiers are represented
by their corresponding key codes (i.e. 0xE0 to 0xE3), and the modifier
is toggled upon encountering the keycodes.