Add rudimentary Unicode support
The full range of Unicode is supported, but the implementation is a
hack. Currently, the system assumes Linux (which is what I use), and
uses the C-S-u combo to get IBus to enter a Unicode character. In order
to reduce complexity, the Unicode characters in the strings are
represented with 4 bytes, a start byte with 0x01 (invalid character/HID
code anyway), and 3 bytes representing the lower 3 bytes of the code
point. It's a hack because the case changing code doesn't take into
account the Unicode formats, and may or may not change the actual
output.