Init
Has a blue LED botched on to `PE6` to ground (through a 1k) indicating
whether it's in embedded or Plover mode (blue is embedded). Also blinks
shortly on power up. Can still use the default `PB0` for entering
dictionary download mode but that's not really reachable.  Uses the
impossible-to-press-in-one-go stroke `#MOEBD` for switching modes.

Built with `b5019b4e` on `crides/qmk_firmware`